requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
crypto==1.4.1
distlib==0.3.6
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
filelock==3.12.0
idna==3.4
mysqlclient==2.2.0
Naked==0.1.32
platformdirs==3.5.0
pycryptodome==3.18.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
shellescape==3.8.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.2
user-agents==2.2.0
virtualenv==20.23.0
